finish list

+++ Platform/CoreData/CoreDataUserCaseProvider.swift
@@ -20,57 +20,69 @@ public final class CoreDataUserCaseProvider: Domain.FavoriteUseCase {
         coreDataStack = CoreDataStack()
     }
     
-//    public func checkFavorite(object: String) -> Observable<Bool> {
-//        <#code#>
-//    }
+    public func checkFavorite(name: String) -> Observable<Bool> {
+        let context =  coreDataStack.context
+        let request = NSFetchRequest<NSFetchRequestResult>(entityName: "CDFavorite")
+        request.predicate = NSPredicate(format: "code = %@",name)
+        do {
+            let result = try context.fetch(request)
+            if result.isEmpty {
+                return Observable.just(false)
+            }else{
+                return Observable.just(true)
+            }
+        } catch {
+            print("Failed")
+            return Observable.just(false)
+        }
+    }
     
     public func fetchFavorite() -> Observable<[Favorite]> {
         let context =  coreDataStack.context
         let request = NSFetchRequest<NSFetchRequestResult>(entityName: "CDFavorite")
+        request.returnsObjectsAsFaults = false
         
         do {
-          let result = try context.fetch(request)
-          return Observable.just(result as? [Favorite] ?? [])
+            let result = try context.fetch(request)
+            
+            var lists = [Favorite]()
+            for data in result as! [CDFavorite] {
+                let content = Favorite(code: data.code ?? "", flagPath: data.flagPath ?? "")
+                lists.append(content)
+            }
+            
+            return Observable.just(lists)
         } catch {
-          print("Failed")
+            print("Failed")
             return Observable.just([])
         }
     }
     
-    public func addFavortie(object: Favorite) -> Observable<Void> {
+    public func triggleFavorite(object: Favorite) -> Bool {
         let context = coreDataStack.context
         context.mergePolicy = NSMergeByPropertyObjectTrumpMergePolicy
-
-        let newFavorite = CDFavorite(context: context)
-        newFavorite.code = object.code
-        newFavorite.flagPath = object.flagPath
         
-        do {
-          try context.save()
-        } catch {
-          print("Failed Add")
-        }
-        return Observable.just(Void())
-    }
-    
-    public func deleteFavorite(object: Favorite) -> Observable<Void> {
-        let context = coreDataStack.context
-        context.mergePolicy = NSMergeByPropertyObjectTrumpMergePolicy
-
         let request = NSFetchRequest<NSFetchRequestResult>(entityName: "CDFavorite")
-        request.predicate = NSPredicate(format: "code == \(object)")
-        
+        request.predicate = NSPredicate(format: "code = %@",object.code)
+        var isFavorite = false
         do {
-          let result = try context.fetch(request)
-          for obj in result as! [NSManagedObject] {
-            context.delete(obj)
-          }
-          
-          try context.save()
+            let result = try context.fetch(request)
+            if result.isEmpty {
+                let newFavorite = CDFavorite(context: context)
+                newFavorite.code = object.code
+                newFavorite.flagPath = object.flagPath
+                isFavorite = true
+            }else{
+                for obj in result as! [NSManagedObject] {
+                    context.delete(obj)
+                }
+            }
+            
+            try context.save()
         } catch {
-          print ("Failed Delete")
+            print ("Failed Triggle")
         }
         
-        return Observable.just(Void())
+        return isFavorite
     }
 }
